Implement Authentication

+++ src/main/java/fpt/practice/moneymanagerment/service/impl/SpendingServiceImpl.java
@@ -1,8 +1,7 @@
 package fpt.practice.moneymanagerment.service.impl;
 
-import fpt.practice.moneymanagerment.constant.StatusCode;
 import fpt.practice.moneymanagerment.dto.SpendingDTO;
-import fpt.practice.moneymanagerment.exception.RestApiException;
+import fpt.practice.moneymanagerment.exception.BadRequestException;
 import fpt.practice.moneymanagerment.model.Account;
 import fpt.practice.moneymanagerment.model.Spending;
 import fpt.practice.moneymanagerment.model.SubSpendingType;
@@ -12,15 +11,15 @@ import fpt.practice.moneymanagerment.repository.SpendingRepository;
 import fpt.practice.moneymanagerment.repository.SubSpendingTypeRepository;
 import fpt.practice.moneymanagerment.repository.UnitRepository;
 import fpt.practice.moneymanagerment.request.SpendingRequest;
+import fpt.practice.moneymanagerment.response.ResponseMessage;
 import fpt.practice.moneymanagerment.service.SpendingService;
-import fpt.practice.moneymanagerment.util.DateUtil;
-import fpt.practice.moneymanagerment.util.NumberUtil;
+import fpt.practice.moneymanagerment.utils.DateUtil;
+import fpt.practice.moneymanagerment.utils.NumberUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -46,32 +45,11 @@ public class SpendingServiceImpl implements SpendingService {
     }
 
     @Override
-    public void addSpending(SpendingRequest spendingRequest) {
+    public void addSpending(SpendingRequest spendingRequest) throws BadRequestException {
         Spending spending = new Spending();
-//        try {
-//            Account account = accountRepository.getById(spendingRequest.getAccountId());
-//            spending.setAccount(account);
-//        }catch (Exception e){
-//            throw new RestApiException(StatusCode.ACCOUNT_NOT_EXIST);
-//        }
-        Optional<Account> account = accountRepository.findById(spendingRequest.getAccountId());
-        if (!account.isPresent()) {
-            throw new RestApiException(StatusCode.ACCOUNT_NOT_EXIST);
-        }
-        spending.setAccount(account.get());
-        try {
-            SubSpendingType subSpendingType = subSpendingTypeRepository.getById(spendingRequest.getSubSpendingTypeId());
-            spending.setSubSpendingType(subSpendingType);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.SUB_SPENDING_TYPE_NOT_EXIST);
-        }
 
-        try {
-            Unit unit = unitRepository.getById(spendingRequest.getUnitId());
-            spending.setUnit(unit);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.UNIT_NOT_EXIST);
-        }
+        spending = checkNotFound(spendingRequest,spending);
+
         spending.setAmount(spendingRequest.getAmount());
         spending.setDescription(spendingRequest.getDescription());
         spending.setDate(DATE_UTIL.getCurrentDate());
@@ -79,47 +57,25 @@ public class SpendingServiceImpl implements SpendingService {
     }
 
     @Override
-    public void updateSpending(Long spendingId, SpendingRequest spendingRequest) {
-        Spending spending = null;
-        try {
-            spending = spendingRepository.getById(spendingId);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.SPENDING_NOT_EXIST);
+    public void updateSpending(Long spendingId, SpendingRequest spendingRequest) throws BadRequestException{
+        Optional<Spending> spendingOptional = spendingRepository.findById(spendingId);
+        if(!spendingOptional.isPresent()){
+            throw new BadRequestException(ResponseMessage.SpendingIdNotFound);
         }
-
-        try {
-            Account account = accountRepository.getById(spendingRequest.getAccountId());
-            spending.setAccount(account);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.ACCOUNT_NOT_EXIST);
-        }
-
-        try {
-            SubSpendingType subSpendingType = subSpendingTypeRepository.getById(spendingRequest.getSubSpendingTypeId());
-            spending.setSubSpendingType(subSpendingType);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.SUB_SPENDING_TYPE_NOT_EXIST);
-        }
-
-        try {
-            Unit unit = unitRepository.getById(spendingRequest.getUnitId());
-            spending.setUnit(unit);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.UNIT_NOT_EXIST);
-        }
+        Spending spending = spendingOptional.get();
+        spending = checkNotFound(spendingRequest,spending);
         spending.setAmount(spendingRequest.getAmount());
         spending.setDescription(spendingRequest.getDescription());
         spendingRepository.save(spending);
     }
 
     @Override
-    public void removeSpending(Long spendingId) {
-        Spending spending = null;
-        try {
-            spending = spendingRepository.getById(spendingId);
-        } catch (Exception e) {
-            throw new RestApiException(StatusCode.SPENDING_NOT_EXIST);
+    public void removeSpending(Long spendingId) throws BadRequestException{
+        Optional<Spending> spendingOptional = spendingRepository.findById(spendingId);
+        if(!spendingOptional.isPresent()){
+            throw new BadRequestException(ResponseMessage.SpendingIdNotFound);
         }
+        Spending spending = spendingOptional.get();
         spendingRepository.delete(spending);
     }
 
@@ -139,5 +95,27 @@ public class SpendingServiceImpl implements SpendingService {
             listSpendingDTOs.add(spendingDTO);
         }
         return listSpendingDTOs;
+    }
+
+    private Spending checkNotFound(SpendingRequest spendingRequest, Spending spending) throws BadRequestException{
+        Optional<Account> account = accountRepository.findById(spendingRequest.getAccountId());
+        if (!account.isPresent()) {
+            throw new BadRequestException(ResponseMessage.AccountIdNotFound);
+        }
+        spending.setAccount(account.get());
+        //check sub spending type
+        Optional<SubSpendingType> subSpendingType = subSpendingTypeRepository.findById(spendingRequest.getSubSpendingTypeId());
+        if(!subSpendingType.isPresent()){
+            throw new BadRequestException(ResponseMessage.SubTypeSpendingIdNotFound);
+        }
+        spending.setSubSpendingType(subSpendingType.get());
+        //check unit
+        Optional<Unit> unit = unitRepository.findById(spendingRequest.getUnitId());
+        if(!unit.isPresent()){
+            throw new BadRequestException(ResponseMessage.UnitIdNotFound);
+        }
+        spending.setUnit(unit.get());
+
+        return spending;
     }
 }
